Implement validation filter branch and slack integration
